The list for rating was supposed to be served by server, but since the Meteor.startUp function did not have Meteor.isServer check, due the the isomorphic nature of functions in Meteor Framework, the data was being inserted into Minimongo again.
So, one possible solution was to add Meteor.isServer() checking while inserting records into the collection for setup.
But, even better was to move such insertion to Server side and hence the insertion is now moved into main.js under server folder so that this
function will not run on the client side any more.